Edited build to only require minio and nginx settings in certain cases. Moved several local loon updates into .build-files
