+++ .build-files/docker-compose-local.template.yml
@@ -62,7 +62,7 @@ services:
     environment:
       DJANGO_ENV_FILE: "/app/.env"
     volumes:
-      - ./.env:/app/.env # relative to docker compose
+      - ${DOCKER_ENV_FILE}:/app/.env # relative to docker compose
     depends_on:
       db:
         condition: service_healthy
@@ -78,7 +78,7 @@ services:
       MYSQL_PASSWORD: ${DATABASE_PASSWORD}
       MYSQL_ROOT_PASSWORD: ${DATABASE_ROOT_PASSWORD}
     volumes:
-      - ./mysql-data:/var/lib/mysql
+      - ${MYSQL_VOLUME_LOCATION}:/var/lib/mysql
     ports:
       - "3306:3306"
     healthcheck:
@@ -96,7 +96,7 @@ services:
       timeout: 20s
       retries: 10
     env_file:
-      - path: ./.env # relative to docker compose (has to be in same directory)
+      - path: ${DOCKER_ENV_FILE} # relative to docker compose (has to be in same directory)
         required: true
 
   redis:
