+++ .build-files/nginx-http-local.conf
@@ -1,0 +1,46 @@
+events {
+    worker_connections 1024;
+}
+
+http {
+    include       mime.types;
+    default_type  application/octet-stream;
+
+    server {
+
+        listen 80;
+
+        client_max_body_size 24000M;
+
+        server_name localhost;
+
+        root /usr/share/nginx/html;
+
+        location / {
+            try_files $uri $uri/ /index.html;
+        }
+
+        location /api {
+            proxy_pass http://server:8000/api;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+
+        location /data {
+            proxy_pass http://data:9000/;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            client_max_body_size 24000M;
+        }
+
+        location ~* \.(ico|css|js|gif|jpe?g|png)$ {
+            # Cache-control headers for static files
+            expires max;
+            add_header Cache-Control "public, must-revalidate, proxy-revalidate";
+        }
+    }
+}
