+++ .gitignore
@@ -59,3 +59,5 @@ __pycache__/
 *.pyc
 config.json
 config-production.json
+config-test.json
+config-local.json
